Update patch set 1
Patch Set 1:

You shouldn't use this task with the CI matrix. This was a way to avoid having to type 5 commands to run a test against a specific nodeset, that's all. Driven by engineering laziness and 'why should I put this in my profile when I can put it in Rake'.

On the STDOUT thing, it was either that or just repeat all of their code. I decided on this. We can change it when they change how their layout works.

I hate whitespace, will patch.

Patch-set: 1
Label: Code-Review=0