(SIMP-1118) Fix `copy_fixture_modules_to` defaults
As a result of BKR-723, `copy_module_to()` defaults to an erroneous 
value in `host['distmoduledir']`.  This has not been updated with a 
valid default for Puppet 4 and breaks against the AIO paths.

This patch addresses the issue by adding a default target_module_path to
`copy_fixture_modules_to()`, which is currently hard-coded to the
default module path for Puppet 4, `/opt/puppetlabs/code/modules.`

As a convenience, the method now also accepts SUTs either as Arrays or a
single-SUT String.

SIMP-1118 #comment hard-coded `:target_module_path` to Puppet 4 defaults

Change-Id: Ic03863a7263c0824bccec708170648e4f4158d9f