(SIMP-1281) Add `#write_hieradata_to` method
This adds a method for fine-grained control of Hiera data.  It is
similar to `#set_hieradata_on`, except it may be used to create multiple
Hiera data files on a system, and it does not manage the Hiera
configuration.  The intent is to allow the creation of somewhat more
realistic Hiera setups with multi-level hierarchies and shared common
data.