(SIMP-1167) Fix for simp-pki compatibility
Presently, we copy in the created CA certificate as cacerts.pem. This
file is expected to be auto-generated by the client and will be ignored
completely by the PKI module making this unsuitable for plugging in on
the fly.

This adds a second cert to the mix that will be properly enabled in the
environment when set as the source in the PKI module.

SIMP-1167 #comment Fix for issue found during testing