(SIMP-3784) Added 'copy_to' functionality
'copy_to' picks the best method for transferring files to the remote
system and performs a copy.

Docker systems can be finicky about rsync connections and using the
native docker commands is MUCH faster so this code will try docker if
appropriate, rsync if available, and fall back to scp.

This does not affect calling the legacy methods.