(SIMP-3831) Fix copy_to rsync call
The upstream version of the rsync call does not behave in the same
manner as either the docker 'tar' behavior or the built-in 'scp'
behavior.

This fixes our call to rsync to work the same way as the other methods
so that all calls to 'copy_to' can expect consistency in behavior.

SIMP-3831 #close