(SIMP-4272) Do not use rsync on FIPS clients
If you are connecting to a FIPS enabled system from a non-FIPS system,
there is a high probability that a rsync connection will fail. Standard
SCP connections are not affected and core Beaker does not provide a
method to pass full options to the sub commands.

SIMP-4272 #comment fips client issue discovered with beaker during testing