(SIMP-6708) Fix inspec report processing (#109)
(SIMP-6708) Fix inspec report processing

The inspec report processor was not processing all sub-controls for a
given test. In the past, this was due to the in-flux nature of the
inspec reports themselves. The latest releases of inspec seem to have
stabilized and can now be easily, and completely processed.
Hard-coded inspec to 4.16.14 by default - 
There is a bug in inspect 4.16.15+ that causes issues with inheritance. See:https://github.com/inspec/inspec/pull/4502

SIMP-6708 #comment fix simp-beaker-helpers bug discovered during testing