Prevent `spec/` directory symlink recursion
Before this patch, the `tar` operation in `copy_fixture_modules_to`
could get stuck on an endless recursive directory loop when the
`spec/fixtures/modules/` directory contained a symlink to the module
itself.

The `spec/` directory is only used for spec testing and is not
required by SUTs to compile the module's code and data.

This patch prevents tar from recursing module fixtures entirely by
adding `spec/` to the directories that `tar` excludes.

Fixes #173