Allow for sequential nodesets
In most modules, we wish to test multiple default nodes but generally
are are not doing multi-node testing.

Instead of using convoluted generators or additional abstract
configuration files, this allows you to simply add `sequential: true` to
the `CONFIG:` section and each host will be broken out into its own
nodeset on the fly and executed.

A future extension of this may allow for tagged clusters to run as
multi-node tests without needing additional files or settings.